src-api/requirements.txt
transformers==4.26.1
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
torch
